@@ -20,8 +20,6 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-using System;
-using System.Threading;
 using System.Threading.Tasks;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
@@ -102,7 +100,7 @@ namespace DSharpPlus.Test
             var five = await ctx.RespondAsync(m => m.WithContent("**Button**: 5/6")
                 .AddComponents(new DiscordButtonComponent(ButtonStyle.Primary, "button-five", "Push me")));
 
-            buttonRes = await ctx.Client.GetInteractivity().WaitForButtonAsync(five, new[] { new DiscordButtonComponent(ButtonStyle.Primary, "button-five", "Push me") } , null);
+            buttonRes = await ctx.Client.GetInteractivity().WaitForButtonAsync(five, new[] { new DiscordButtonComponent(ButtonStyle.Primary, "button-five", "Push me") }, null);
 
             if (!buttonRes.TimedOut)
             {
