@@ -31,14 +31,14 @@ namespace DSharpPlus.Test
 {
     public class PaginationTest : BaseCommandModule
     {
-        private const string _lorem = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce mattis turpis dapibus risus bibendum condimentum. Mauris ut dictum arcu. Nullam id ante nisl. Morbi lobortis nisi dignissim nisl pulvinar, eu suscipit augue pellentesque. Sed vehicula quam magna, id lobortis diam condimentum sed. Sed in mi felis. Cras sodales dui purus, sed gravida nulla venenatis ac. Aliquam venenatis pharetra tellus, eget pretium felis ultricies ac. Donec a tristique lacus, ac vulputate tellus. Praesent condimentum fringilla venenatis. Nam quis dui ut ante scelerisque scelerisque ac dictum sem. Donec non tristique ex. Quisque iaculis felis non tincidunt finibus. Praesent pellentesque sit amet tortor sit amet viverra. Aenean vestibulum est sit amet mauris faucibus sodales. Praesent tristique lacus at lorem consectetur, ut accumsan lorem tincidunt." +
+        private const string Lorem = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce mattis turpis dapibus risus bibendum condimentum. Mauris ut dictum arcu. Nullam id ante nisl. Morbi lobortis nisi dignissim nisl pulvinar, eu suscipit augue pellentesque. Sed vehicula quam magna, id lobortis diam condimentum sed. Sed in mi felis. Cras sodales dui purus, sed gravida nulla venenatis ac. Aliquam venenatis pharetra tellus, eget pretium felis ultricies ac. Donec a tristique lacus, ac vulputate tellus. Praesent condimentum fringilla venenatis. Nam quis dui ut ante scelerisque scelerisque ac dictum sem. Donec non tristique ex. Quisque iaculis felis non tincidunt finibus. Praesent pellentesque sit amet tortor sit amet viverra. Aenean vestibulum est sit amet mauris faucibus sodales. Praesent tristique lacus at lorem consectetur, ut accumsan lorem tincidunt." +
                                       "Pellentesque ipsum magna, laoreet eu erat id, sagittis posuere nulla. Nulla suscipit dui luctus lorem consectetur, et hendrerit est cursus. Integer malesuada, mi et viverra sagittis, urna enim finibus eros, sed vestibulum ipsum leo ut massa. In lacinia risus in commodo pretium. Interdum et malesuada fames ac ante ipsum primis in faucibus. Donec eu pulvinar dolor. Nullam molestie, lectus at tincidunt laoreet, nunc magna accumsan nibh, ac faucibus erat enim vel lorem." +
                                       "" +
                                       "" +
                                       "";
 
-        [Command]
-        public async Task Paginate(CommandContext ctx)
+        [Command("paginate")]
+        public async Task PaginateAsync(CommandContext ctx)
         {
             var builder = new DiscordMessageBuilder().WithContent("** **").AddComponents(new DiscordButtonComponent(ButtonStyle.Primary, "a", "Paginate"));
             await ctx.RespondAsync(builder);
@@ -46,13 +46,14 @@ namespace DSharpPlus.Test
             ctx.Client.ComponentInteractionCreated += this.Handle;
         }
 
-        private async Task Handle(DiscordClient sender, ComponentInteractionCreateEventArgs e)
+        private Task Handle(DiscordClient sender, ComponentInteractionCreateEventArgs e)
         {
             if (e.Id != "a")
-                return;
-            var pages = sender.GetInteractivity().GeneratePagesInContent(_lorem);
+                return Task.CompletedTask;
+            var pages = sender.GetInteractivity().GeneratePagesInContent(Lorem);
             _ = sender.GetInteractivity().SendPaginatedResponseAsync(e.Interaction, true, e.User, pages);
             sender.ComponentInteractionCreated -= this.Handle;
+            return Task.CompletedTask;
         }
     }
 }
