@@ -1,4 +1,4 @@
-﻿// This file is part of the DSharpPlus project.
+// This file is part of the DSharpPlus project.
 //
 // Copyright (c) 2015 Mike Santiago
 // Copyright (c) 2016-2022 DSharpPlus Contributors
@@ -30,9 +30,7 @@ namespace DSharpPlus.Test
     public class SlashCommandMinMaxTest : ApplicationCommandModule
     {
         [SlashCommand("minmax", "testing min and max")]
-        public async Task MinMax(InteractionContext ctx, [Option("value", "value with limits")] [Minimum(5), Maximum(7)] long thing)
-        {
-            await ctx.CreateResponseAsync(InteractionResponseType.ChannelMessageWithSource, new DiscordInteractionResponseBuilder().WithContent(thing.ToString()));
-        }
+        public async Task MinMaxAsync(InteractionContext ctx, [Option("value", "value with limits")][Minimum(5), Maximum(7)] long thing)
+            => await ctx.CreateResponseAsync(InteractionResponseType.ChannelMessageWithSource, new DiscordInteractionResponseBuilder().WithContent(thing.ToString()));
     }
 }
