@@ -97,7 +97,7 @@ namespace DSharpPlus.Test
         }
 
         [Command("addsimple"), Description("Adds a simple echo command."), Hidden, RequireOwner]
-        public async Task AddSimple(CommandContext ctx, string name, params string[] aliases)
+        public async Task AddSimpleAsync(CommandContext ctx, string name, params string[] aliases)
         {
             await Task.Yield();
             var command = new CommandBuilder(null)
