@@ -25,28 +25,32 @@ using System.Threading.Tasks;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
 using DSharpPlus.VoiceNext;
-using Microsoft.Extensions.Logging;
 
 namespace DSharpPlus.Test
 {
     public class VoiceNextTest : BaseCommandModule
     {
-        static VoiceNextTest() => TaskScheduler.UnobservedTaskException += OhNo;
-        private static void OhNo(object sender, UnobservedTaskExceptionEventArgs e) { Console.Error.WriteLine("SOMETHING WENT TERRIBLY WRONG WHEN DISCONNECTING"); }
+        static VoiceNextTest()
+        {
+            TaskScheduler.UnobservedTaskException += OhNo;
+        }
+
+        private static void OhNo(object sender, UnobservedTaskExceptionEventArgs e) => Console.Error.WriteLine("SOMETHING WENT TERRIBLY WRONG WHEN DISCONNECTING");
 
         [Command]
-        public async Task Join(CommandContext ctx)
+        public async Task JoinAsync(CommandContext ctx)
         {
             var vnext = ctx.Client.GetVoiceNext();
             await vnext.ConnectAsync(ctx.Member.VoiceState.Channel);
         }
 
         [Command]
-        public async Task Leave(CommandContext ctx)
+        public static Task Leave(CommandContext ctx)
         {
             var vnext = ctx.Client.GetVoiceNext();
 
             vnext.GetConnection(ctx.Guild)?.Disconnect(); // Calls .Dispose(); //
+            return Task.CompletedTask;
         }
     }
 }
